requirements.txt
annotated-types==0.7.0
anyio==4.12.0
asgiref==3.11.0
cachetools==6.2.2
certifi==2025.11.12
charset-normalizer==3.4.4
distro==1.9.0
Django==5.0
djangorestframework==3.14.0
google-ai-generativelanguage==0.6.4
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.43.0
google-auth-httplib2==0.2.1
google-generativeai==0.6.0
googleapis-common-protos==1.72.0
grpcio==1.76.0
grpcio-status==1.62.3
gunicorn==21.2.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
idna==3.11
lxml==6.0.2
openai==1.6.1
packaging==25.0
proto-plus==1.26.1
protobuf==4.25.8
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.12.5
pydantic_core==2.41.5
pyparsing==3.2.5
PyPDF2==3.0.1
python-docx==1.1.0
python-dotenv==1.0.0
pytz==2025.2
requests==2.31.0
rsa==4.9.1
sniffio==1.3.1
sqlparse==0.5.4
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.6.0
whitenoise==6.6.0
